RAAS/requirements-tux.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
Cython==0.29.33
defusedxml==0.7.1
discord==2.1.0
discord.py==2.1.0
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.4.0/fr_core_news_sm-3.4.0-py3-none-any.whl
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
joblib==1.1.1
langcodes==3.3.0
MarkupSafe==2.1.2
mlconjug3==3.8.3
multidict==6.0.4
murmurhash==1.0.9
numpy==1.22.4
packaging==23.0
pathy==0.10.1
preshed==3.0.8
progressbar2==3.53.3
pydantic==1.10.4
python-dotenv==0.21.0
python-utils==3.4.5
requests==2.28.2
scikit-learn==1.0.2
scipy==1.10.0
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
threadpoolctl==3.1.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==0.10.1
yarl==1.8.2
